+++ interface_stone_smart_flutter/lib/utils/types/payment_type_call.dart
@@ -21,4 +21,6 @@ enum PaymentTypeCall {
   printFromBase64,
   printWrapPaper,
   activeDebugLog,
+  getPosSerialNumber,
+  getPosManufacture,
 }